refactor: 파일 저장 구현체 파일 옮김
